Port tcConv's v1 parsing algorithm because it's much more faster
Enable v2 timecode outputting again, now as an option.
Both cfr and v1 parsing can output v2 timecodes to a file.
Finally, update tccreator.py and rename it to tcconv.py, as it does
functionally the same (v1 to v2) and also fps to v2. No v2 to v1 as it's useless.