Remove deprecated use of % operator, fix incorrect use of subprocess.call and subprocess.check_output
You’re supposed to pass a list of parameters to subprocess.call,
not a single string containing a shell command.

test.py was incorrectly using 'python' to execute vfr.py rather than
'python3'. It also used a backslash as path separator, which is not
portable. Fixed.
This commit slightly changes some output of vfr.py, so right now
test.py will assume that the commands failed.